Add files via upload
This script computes interpolated fields of p-values for pairwise significance of difference tests.
The objective is to determine whether integration between specified pairs of bones changes significantly as a function of body mass.
The interpolated fields are defined such that the x and y axes represent the mean masses of heavier and lighter subsamples of birds. 
Therefore the distribution of p-values across this space illustrates significant changes in integration with differences in body mass.